requirements.txt
amqp==5.1.1
asgiref==3.5.1
asn1crypto==0.24.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==2.1.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
executing==0.8.3
idna==2.6
ipython==8.3.0
jedi==0.18.1
keyring==10.6.0
keyrings.alt==3.0
kombu==5.2.4
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycrypto==2.6.1
Pygments==2.12.0
PyGObject==3.26.1
pyparsing==3.0.8
pytz==2022.1
pyxdg==0.25
redis==4.2.2
requests==2.27.1
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
